Packer - tweaks to get everything compat w/ new kalilinux-kali-rolling base container

+++ build/provisioners/upload_globals.sh
@@ -1,8 +1,11 @@
 #!/bin/bash --login
 cat >/etc/profile.d/globals.sh <<'EOF'
 #!/bin/bash --login
+export DEBIAN_FRONTEND=noninteractive
+export TERM=xterm
+
 screen_session=`basename ${0} .sh`
-screen_cmd="screen -L -S ${screen_session} -d -m /bin/bash --login -c"
+screen_cmd="screen -T xterm -L -S ${screen_session} -d -m /bin/bash --login -c"
 assess_update_errors='|| echo IMAGE_ABORT && exit 1'
 debconf_set='/usr/bin/debconf-set-selections'
 apt="DEBIAN_FRONTEND=noninteractive apt -o Dpkg::Options::='--force-confdef' -o Dpkg::Options::='--force-confnew'"
@@ -21,7 +24,10 @@ grok_error() {
         exit 1
       else
         echo "No errors in $(ls /screenlog.*) detected...moving onto the next."
-        rm /screenlog.*
+        ls /screenlog.* > /dev/null 2>&1
+        if [[ $? == 0 ]]; then
+          rm /screenlog.*
+        fi
         break
       fi
     else
